requirements-rag.txt
# Orchestration & eval
langgraph==0.6.8
langgraph-checkpoint-redis==0.1.2
langchain==0.3.27
langchain-openai==0.3.35
python-dotenv==1.1.1
tenacity==9.1.2
tiktoken==0.12.0

# Vector DB & hybrid search
qdrant-client==1.15.1
bm25s[hf]==0.2.14
PyStemmer==3.0.0

# Chunking / parsing
tree_sitter_languages==1.10.2

# Reranking & Torch (MPS on Apple Silicon)
sentence-transformers==5.1.1
transformers==4.57.0
accelerate==1.10.1
torch==2.8.0

# Service
fastapi==0.118.0
uvicorn==0.37.0
redis==6.4.0
pydantic==2.11.10
openai==2.2.0
